requirements.txt
#required libraries
langchain
langchain-core
langchain-groq
pydantic
python-dotenv
streamlit